requirements.txt
streamlit==1.29.0
pandas==2.2.2
google-generativeai==0.2.2  
python-dotenv
fpdf
python-docx
pdfplumber
PyPDF2
pillow
pytesseract